requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==20.3.0
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.0.5
django-timezone-field==4.1.2
djangorestframework==3.12.4
iniconfig==1.1.1
kombu==5.0.2
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
